refactor: reorganize domain layer by aggregate (colocation pattern)
- Move from technical separation (entities/, repositories/) to aggregate-based structure
- Each aggregate (manufacturer, category, product, asset) now has its own directory
- Value objects colocated with their aggregates:
  - TransactionType -> asset/transaction_type.rs
  - MeasurementUnit -> product/measurement_unit.rs
- Benefits:
  - Clear aggregate boundaries (DDD)
  - High locality of change
  - Better encapsulation
  - Easier to navigate and maintain

Structure:
domain/
├── manufacturer/     # Manufacturer aggregate
├── category/         # Category aggregate
├── product/          # Product aggregate (+ MeasurementUnit VO)
├── asset/            # Asset aggregate (+ AssetTransaction + TransactionType VO)
└── errors.rs         # Shared domain errors
