--- src-tauri/contexts/asset/src/domain.rs
@@ -1,5 +1,0 @@
-// Asset Context - Domain Layer
-
-pub mod entities;
-pub mod errors;
-pub mod repositories;
