+++ src-tauri/contexts/asset/src/domain/asset/mod.rs
@@ -1,0 +1,14 @@
+// Asset Aggregate
+
+pub mod asset;
+pub mod repository;
+pub mod transaction;
+pub mod transaction_repository;
+pub mod transaction_type;
+
+// Re-exports
+pub use asset::Asset;
+pub use repository::AssetRepository;
+pub use transaction::AssetTransaction;
+pub use transaction_repository::AssetTransactionRepository;
+pub use transaction_type::TransactionType;
