+++ src-tauri/contexts/asset/src/domain/category/entity.rs
