+++ src-tauri/contexts/asset/src/domain/category/mod.rs
@@ -1,0 +1,8 @@
+// Category Aggregate
+
+pub mod entity;
+pub mod repository;
+
+// Re-exports
+pub use entity::Category;
+pub use repository::CategoryRepository;
