+++ src-tauri/contexts/asset/src/domain/manufacturer/entity.rs
