+++ src-tauri/contexts/asset/src/domain/manufacturer/mod.rs
@@ -1,0 +1,8 @@
+// Manufacturer Aggregate
+
+pub mod entity;
+pub mod repository;
+
+// Re-exports
+pub use entity::Manufacturer;
+pub use repository::ManufacturerRepository;
