+++ src-tauri/contexts/asset/src/domain/mod.rs
@@ -1,0 +1,10 @@
+// Asset Context - Domain Layer
+
+// Aggregates (集約単位でコロケーション)
+pub mod asset;
+pub mod category;
+pub mod manufacturer;
+pub mod product;
+
+// Shared
+pub mod errors;
