+++ src-tauri/contexts/asset/src/domain/product/mod.rs
@@ -1,0 +1,10 @@
+// Product Aggregate
+
+pub mod entity;
+pub mod measurement_unit;
+pub mod repository;
+
+// Re-exports
+pub use entity::Product;
+pub use measurement_unit::MeasurementUnit;
+pub use repository::ProductRepository;
