--- src-tauri/contexts/asset/src/domain/repositories.rs
@@ -1,7 +1,0 @@
-// Asset Context - Repository Interfaces
-
-pub mod asset;
-pub mod asset_transaction;
-pub mod category;
-pub mod manufacturer;
-pub mod product;
