+++ src-tauri/contexts/asset/src/lib.rs
@@ -9,6 +9,5 @@ pub mod presentation;
 pub use application::dto;
 pub use application::errors::ApplicationError;
 pub use application::services;
-pub use domain::entities;
 pub use domain::errors::DomainError;
 pub use presentation::graphql;
